raiden: transport: move _handle_sdp_callback to WebRTCManager
